Warn on locale fallback, and fix agreement in six catalogues
**Observability.** A run that asked for a locale and got English said so
only at `DEBUG`, which the startup filter never reaches. That is exactly
the case someone would report as a bug, so it now warns: both for a tag
that will not parse and for one that ships no catalogue and whose
language ships none. A resolution that lands on a real catalogue stays
at `DEBUG`, and a shipped locale warns about nothing, or the warning
would fire on every ordinary run and stop meaning anything. Tests assert
the level rather than the event, since the level is what decides
visibility; reverting the unsupported path to `DEBUG` turns the `is-IS`
case red.

**Agreement and terminology.** Czech `stdlib.command.*` agreed a neuter
participle with `{ $location }`, whose value is headed by the masculine
`příkaz`. Italian agreed a feminine participle with `{ $action }`, whose
values are infinitives; the four templates now make `operazione` the
subject. Norwegian's ten action values were infinitives, so appending
`mislyktes` was ungrammatical, and they become noun phrases as the
Danish ones did. European Portuguese described flushing as emptying the
file and stdout rather than their buffers, where every other catalogue
including pt-BR names the buffer. Romanian used `Marcați` — "mark" — for
escaping. Turkish used `çevirir` where the rest of that catalogue uses
`derlemek` for compilation.

The German claim in the same review is declined: `übersetzen` is the
standard German term for compiling — a compiler is an *Übersetzer* — and
cs, da, and fi use the same translate-rooted verb.

British English had five Americanisms in prose whose keys were already
British, plus `canonicalize` and `Synthesizing`. Only the message values
change: the keys are identifiers shared across catalogues, and renaming
three of them broke key parity until the audit caught it.

**File length.** `AGENTS.md` caps a code file at 400 lines;
`localization_tests.rs` had reached 648. The plural suite moves to
`localization_plural_tests.rs`, which is where it belongs anyway — it
tests what Fluent selects, where `locale_catalogue_tests.rs` tests what
the catalogues declare. `localizer_guards` moves into `test_support`
rather than being duplicated.

**Two test defects.** The registry's duplicate check compared the
registry's size against the expected list's length, which measures the
registry rather than the list its message names. And the plural oracle
took its default category from the first `*[` in the whole catalogue,
which belongs to whichever select comes first; it now comes from the
requested key's own variants.
