+++ locales/cs/messages.ftl
@@ -221,13 +221,13 @@ stdlib.fetch.action.open_cache_entry = otevření položky mezipaměti fetch
 # Diagnostika pomocníka pro příkazy.
 stdlib.command.location = příkaz „{ $command }“ v šabloně „{ $template }“
 stdlib.command.spawn_failed = { $location } se nepodařilo spustit: { $details }.
-stdlib.command.io_failed = { $location } selhalo: { $details }.
+stdlib.command.io_failed = { $location } selhal: { $details }.
 stdlib.command.closed_input_early = Vstup se uzavřel dříve, než byl zápis do příkazu dokončen.
 stdlib.command.broken_pipe = Přerušená roura při běhu { $location }: { $details }.
-stdlib.command.terminated_by_signal = { $location } bylo ukončeno signálem.
-stdlib.command.exited_with_status = { $location } skončilo se stavem { $status }.
-stdlib.command.output_limit_exceeded = { $location } překročilo limit { $mode } ve výši { $limit } bajtů pro { $stream }.
-stdlib.command.timeout = { $location } překročilo časový limit { $seconds } s.
+stdlib.command.terminated_by_signal = { $location } byl ukončen signálem.
+stdlib.command.exited_with_status = { $location } skončil se stavem { $status }.
+stdlib.command.output_limit_exceeded = { $location } překročil limit { $mode } ve výši { $limit } bajtů pro { $stream }.
+stdlib.command.timeout = { $location } překročil časový limit { $seconds } s.
 stdlib.command.exit_status_suffix = (návratový kód { $status })
 stdlib.command.signal_suffix = (ukončeno signálem)
 stdlib.command.shell.empty = Příkaz shellu nesmí být prázdný.
