+++ locales/en-GB/messages.ftl
@@ -55,7 +55,7 @@ cli.validation.color.invalid = Invalid colour policy '{ $value }'. Valid options
 cli.validation.emoji.invalid = Invalid emoji policy '{ $value }'. Valid options: auto, always, never.
 cli.validation.progress.invalid = Invalid progress policy '{ $value }'. Valid options: auto, always, never.
 cli.validation.accessibility.invalid = Invalid accessibility policy '{ $value }'. Valid options: auto, on, off.
-cli.validation.config.expected_object = Expected parsed CLI values to serialize to an object, got { $value }.
+cli.validation.config.expected_object = Expected parsed CLI values to serialise to an object, got { $value }.
 
 # Clap error messages.
 clap-error-missing-argument = Missing required argument: { $argument }
@@ -78,7 +78,7 @@ runner.manifest.directory_label = directory `{ $directory }`
 runner.manifest.current_directory_label = the current directory
 runner.context.network_policy = Failed to build the network policy.
 runner.context.load_manifest = Failed to load manifest at { $path }.
-runner.context.serialise_manifest = Failed to serialize manifest.
+runner.context.serialise_manifest = Failed to serialise manifest.
 runner.context.build_graph = Failed to build graph from the manifest.
 runner.context.generate_ninja = Failed to generate the Ninja manifest.
 runner.context.render_graph = Failed to render the graph artefact.
@@ -120,7 +120,7 @@ manifest.path_non_utf8 = Manifest '{ $manifest }' path is not valid UTF-8: { $pa
 manifest.path_missing_name = Manifest path '{ $path }' has no file name.
 manifest.open_workspace_failed = Failed to open workspace { $workspace } for manifest { $manifest }.
 manifest.foreach.not_iterable = `foreach` expression is not iterable.
-manifest.foreach.serialise_item = Failed to serialize foreach item.
+manifest.foreach.serialise_item = Failed to serialise foreach item.
 manifest.when.empty = `when` expression must not be empty.
 manifest.when.eval_error = Failed to evaluate `when` expression '{ $expr }'.
 manifest.when.template_error = Failed to render `when` template '{ $expr }'.
@@ -155,7 +155,7 @@ ir.multiple_rules = Target '{ $target }' must reference a single rule, got { $ru
 ir.empty_rule = Target '{ $target }' must reference a rule.
 ir.duplicate_outputs = Duplicate outputs detected: { $outputs }.
 ir.circular_dependency = Circular dependency detected: { $cycle }.
-ir.action_serialisation = Failed to serialize action: { $details }.
+ir.action_serialisation = Failed to serialise action: { $details }.
 ir.invalid_command = Invalid command interpolation: { $snippet }.
 
 # Ninja generation errors.
@@ -285,7 +285,7 @@ stdlib.path.io.network_down = network down
 stdlib.path.io.network_unreachable = network unreachable
 stdlib.path.io.host_unreachable = host unreachable
 stdlib.path.io.other = I/O error
-stdlib.path.action.canonicalize = canonicalize
+stdlib.path.action.canonicalize = canonicalise
 stdlib.path.action.open_directory = open directory
 stdlib.path.action.stat = stat
 stdlib.path.action.read = read
@@ -327,7 +327,7 @@ stdlib.which.command.empty = which requires a non-empty string.
 stdlib.which.cwd_mode.invalid = cwd_mode must be 'auto', 'always', or 'never', got '{ $mode }'.
 stdlib.which.cwd.resolve_failed = Failed to resolve current directory: { $details }.
 stdlib.which.cwd.non_utf8 = Current directory contains non-UTF-8 components.
-stdlib.which.canonicalize_failed = Failed to canonicalize '{ $path }': { $details }.
+stdlib.which.canonicalize_failed = Failed to canonicalise '{ $path }': { $details }.
 stdlib.which.is_executable = Failed to inspect whether '{ $path }' is executable: { $details }.
 stdlib.which.canonicalize_non_utf8 = Canonical path contains non-UTF-8 components.
 stdlib.which.workspace_non_utf8 = Workspace path contains non-UTF-8 components while resolving command '{ $command }': { $path }.
@@ -351,10 +351,10 @@ status.task.progress_update = { $task }: { $description }
 status.stage.manifest_ingestion = Reading manifest file
 status.stage.initial_yaml_parsing = Parsing YAML document
 status.stage.template_expansion = Expanding template directives
-status.stage.final_rendering = Deserializing and rendering manifest values
+status.stage.final_rendering = Deserialising and rendering manifest values
 status.stage.ir_generation_validation = Building and validating dependency graph
-status.stage.ninja_synthesis = Synthesizing Ninja build plan
-status.stage.ninja_synthesis_execute = Synthesizing Ninja plan and executing { $tool }
+status.stage.ninja_synthesis = Synthesising Ninja build plan
+status.stage.ninja_synthesis_execute = Synthesising Ninja plan and executing { $tool }
 status.stage.graph_rendering = Rendering graph artefact
 status.stage.graph_rendering_with_tool = Rendering { $tool }
 status.complete = { $tool } complete.
