+++ locales/nb/messages.ftl
@@ -212,11 +212,11 @@ stdlib.fetch.cache_write_failed = Hurtiglageret for «{ $url }» kunne ikke skri
 stdlib.fetch.response_limit_exceeded = Svaret fra «{ $url }» oversteg grensen på { $limit } byte.
 stdlib.fetch.cache_limit_exceeded = Det hurtiglagrede svaret «{ $name }» oversteg grensen på { $limit } byte.
 stdlib.fetch.io_failed = { $action } mislyktes for { $path }: { $details }.
-stdlib.fetch.action.sync_cache = synkronisere fetch-hurtiglageret
-stdlib.fetch.action.create_cache_dir = opprette katalogen for fetch-hurtiglageret
-stdlib.fetch.action.open_cache_dir = åpne katalogen for fetch-hurtiglageret
-stdlib.fetch.action.stat_cache = slå opp oppføringen i fetch-hurtiglageret
-stdlib.fetch.action.open_cache_entry = åpne oppføringen i fetch-hurtiglageret
+stdlib.fetch.action.sync_cache = synkronisering av fetch-hurtiglageret
+stdlib.fetch.action.create_cache_dir = oppretting av katalogen for fetch-hurtiglageret
+stdlib.fetch.action.open_cache_dir = åpning av katalogen for fetch-hurtiglageret
+stdlib.fetch.action.stat_cache = oppslag på oppføringen i fetch-hurtiglageret
+stdlib.fetch.action.open_cache_entry = åpning av oppføringen i fetch-hurtiglageret
 
 # Diagnostikk for kommandohjelperen.
 stdlib.command.location = kommandoen «{ $command }» i malen «{ $template }»
@@ -285,11 +285,11 @@ stdlib.path.io.network_down = nettverket er nede
 stdlib.path.io.network_unreachable = nettverket kan ikke nås
 stdlib.path.io.host_unreachable = verten kan ikke nås
 stdlib.path.io.other = I/U-feil
-stdlib.path.action.canonicalize = kanonisere
-stdlib.path.action.open_directory = åpne katalog
-stdlib.path.action.stat = slå opp
-stdlib.path.action.read = lese
-stdlib.path.action.open_file = åpne fil
+stdlib.path.action.canonicalize = kanonisering
+stdlib.path.action.open_directory = åpning av katalog
+stdlib.path.action.stat = oppslag
+stdlib.path.action.read = lesing
+stdlib.path.action.open_file = åpning av fil
 stdlib.path.with_suffix.empty_separator = with_suffix krever et skilletegn som ikke er tomt.
 stdlib.path.relative_to.mismatch = { $path } er ikke relativ til { $root }.
 stdlib.path.expanduser.unsupported = Brukerspesifikk utvidelse av ~ støttes ikke.
