+++ locales/ro/messages.ftl
@@ -109,7 +109,7 @@ manifest.yaml.hint.list_item = Elementele de listă YAML trebuie să înceapă c
 manifest.yaml.hint.expected_colon = Pare o intrare de mapare; lipsește „:” după cheie.
 manifest.yaml.hint.mapping_values = Mapările YAML cer o valoare după „:” (sau un bloc imbricat).
 manifest.yaml.hint.invalid_token = Simbolul YAML este nevalid sau neașteptat.
-manifest.yaml.hint.escape = Marcați barele oblice inverse sau eliminați secvențele de evadare nevalide.
+manifest.yaml.hint.escape = Aplicați escape barelor oblice inverse sau eliminați secvențele de escape nevalide.
 manifest.env.missing = Variabila de mediu obligatorie „{ $name }” nu este setată.
 manifest.env.invalid_utf8 = Variabila de mediu „{ $name }” conține UTF-8 nevalid.
 manifest.vars.not_object = Câmpul `vars` al manifestului trebuie să fie o mapare sau un obiect.
