+++ locales/tr/messages.ftl
@@ -1,6 +1,6 @@
 # Netsuke komut satırı için yerelleştirme kaynakları.
 
-cli.about = Netsuke, YAML + Jinja bildirimlerini Ninja derleme planlarına çevirir.
+cli.about = Netsuke, YAML + Jinja bildirimlerini Ninja derleme planlarına derler.
 cli.long_about = Netsuke, YAML + Jinja bildirimlerini yeniden üretilebilir Ninja çizgelerine dönüştürür ve Ninja'yı güvenli varsayılanlarla çalıştırır.
 cli.usage = { $usage }
